+++ app/src/info/guardianproject/mrapp/publish/RenderWorker.java
@@ -56,7 +56,7 @@ public class RenderWorker extends WorkerBase {
 	}
 
     @Override
-    public void jobProgress(Job job, int progress, String message) {
+    public void jobProgress(Job job, float progress, String message) {
         mController.jobProgress(job, progress, message);
         
     }
