+++ app/src/info/guardianproject/mrapp/publish/sites/FacebookPublisher.java
@@ -53,7 +53,7 @@ public class FacebookPublisher extends PublisherBase {
 		// nop
 	}
 	
-	public void jobProgress(Job job, int progress, String message) {
+	public void jobProgress(Job job, float progress, String message) {
 	    mController.publishJobProgress(mPublishJob, progress, message);
 	}
 }
