+++ app/src/info/guardianproject/mrapp/publish/sites/FlickrPublisher.java
@@ -69,7 +69,7 @@ public class FlickrPublisher extends PublisherBase
     }
 
     @Override
-    public void jobProgress(Job job, int progress, String message) 
+    public void jobProgress(Job job, float progress, String message) 
     {
         Log.d(TAG, "jobProgress()");
